+++ script/tool/test/list_command_test.dart
@@ -101,15 +101,18 @@ void main() {
           '/packages/plugin1/pubspec.yaml',
           '/packages/plugin1/AUTHORS',
           '/packages/plugin1/CHANGELOG.md',
+          '/packages/plugin1/README.md',
           '/packages/plugin1/example/pubspec.yaml',
           '/packages/plugin2/pubspec.yaml',
           '/packages/plugin2/AUTHORS',
           '/packages/plugin2/CHANGELOG.md',
+          '/packages/plugin2/README.md',
           '/packages/plugin2/example/example1/pubspec.yaml',
           '/packages/plugin2/example/example2/pubspec.yaml',
           '/packages/plugin3/pubspec.yaml',
           '/packages/plugin3/AUTHORS',
           '/packages/plugin3/CHANGELOG.md',
+          '/packages/plugin3/README.md',
         ]),
       );
     });